requirements.txt
pyimgui-wheels==2.0.1
tensorflow==2.8.0
protobuf==3.20.0
glfw==2.5.5
PyOpenGL==3.1.6
numpy>=1.23.2
mrcfile>=1.4.3
Pillow>=9.2.0
scipy>=1.9.1
tifffile>=2022.8.12
dill>=0.3.5.1